fix loi migrate khong chay duoc

+++ database/migrations/2024_09_25_072549_create_promotions_table.php
@@ -16,12 +16,12 @@ return new class extends Migration
             $table->string('code')->unique();
             $table->string('description');
             $table->tinyInteger('discount_type')->default(1)->comment('1: percent, 2: amount');
-            $table->bigInteger('discount_value');
+            $table->integer('discount_value');
             $table->dateTime('start_date');
             $table->dateTime('end_date');
             $table->integer('quantity');
-            $table->bigInteger('min_order_total')->nullable();
-            $table->bigInteger('max_discount')->nullable();
+            $table->integer('min_order_total')->nullable();
+            $table->integer('max_discount')->nullable();
             $table->tinyInteger('is_global')->default(2)->comment('1: global, 2: not global');
             $table->tinyInteger('status')->default(1)->comment('1: active, 2: inactive');
             $table->timestamps();
